some more progress.

+++ src/crispy/range.h
@@ -34,7 +34,7 @@ struct range
     constexpr const_iterator cbegin() const { return begin_; }
     constexpr const_iterator cend() const { return end_; }
 
-    constexpr size_t size() const noexcept { return std::distance(begin_, end_); }
+    constexpr size_t size() const noexcept { return static_cast<size_t>(std::distance(begin_, end_)); }
 };
 
 template <typename Iter>
